seperate evaluator and comparer
