+++ medmodels/statistic_evaluations/evaluate_compare/compare.pyi
@@ -1,42 +1,15 @@
 from __future__ import annotations
 
-from typing import Dict, List, Optional, Tuple, TypedDict, Union
+from typing import Dict, List, Tuple, TypedDict
 
-from medmodels.medrecord.medrecord import MedRecord
-from medmodels.medrecord.querying import NodeOperation
 from medmodels.medrecord.types import (
     AttributeInfo,
     AttributeSummary,
     Group,
-    GroupInputList,
     MedRecordAttribute,
     NodeIndex,
 )
-
-class Cohort:
-    """Configuration for a cohort for evaluation and comparison.
-
-    Needs a medrecord and the corresponding patient group that should form the cohort.
-    The cohort group can be a predefined group in the MedRecord or a node query.
-
-    """
-
-    medrecord: MedRecord
-    name: str
-    cohort_group: Group
-    time_attribute: MedRecordAttribute
-    attributes: Optional[Dict[str, MedRecordAttribute]]
-    concepts_groups: Optional[GroupInputList]
-
-    def __init__(
-        self,
-        medrecord: MedRecord,
-        name: str,
-        cohort_group: Union[Group, NodeOperation] = "patients",
-        time_attribute: MedRecordAttribute = "time",
-        attributes: Optional[Dict[str, MedRecordAttribute]] = None,
-        concepts_groups: Optional[GroupInputList] = None,
-    ) -> None: ...
+from medmodels.statistic_evaluations.evaluate_compare.evaluate import CohortEvaluator
 
 class CohortSummary(TypedDict):
     """Dictionary for the cohort summary."""
@@ -64,52 +37,45 @@ class TestSummary(TypedDict):
     not_reject: bool
     p_value: float
 
-class DataComparer:
+class CohortComparer:
     @staticmethod
     def compare_cohort_attribute(
-        cohorts_attribute: List[Tuple[Cohort, MedRecordAttribute]],
+        cohorts: List[CohortEvaluator],
+        attribute: MedRecordAttribute,
     ) -> Dict[str, AttributeInfo]: ...
     @staticmethod
     def test_difference_attribute(
-        cohorts_attribute: List[Tuple[Cohort, MedRecordAttribute]],
+        cohorts_attribute: List[CohortEvaluator],
+        attribute: MedRecordAttribute,
         significance_level: float,
     ) -> List[TestSummary]: ...
     @staticmethod
     def compare_cohorts(
-        cohorts: List[Cohort],
+        cohorts: List[CohortEvaluator],
     ) -> Dict[str, CohortSummary]: ...
     @staticmethod
     def test_difference_cohort_attributes(
-        cohorts: List[Cohort],
+        cohorts: List[CohortEvaluator],
         significance_level: float,
     ) -> Dict[str, List[TestSummary]]: ...
     @staticmethod
     def calculate_absolute_relative_difference(
-        control_group: Cohort,
-        case_groups: List[Cohort],
+        control_group: CohortEvaluator,
+        case_group: CohortEvaluator,
     ) -> Tuple[float, Dict[MedRecordAttribute, float]]: ...
     @staticmethod
-    def get_concept_counts(
-        cohort: Cohort,
-    ) -> List[Tuple[NodeIndex, int]]: ...
-    @staticmethod
-    def get_top_k_concepts(
-        cohort: Cohort,
-        top_k: int,
-    ) -> List[NodeIndex]: ...
-    @staticmethod
     def test_difference_top_k_concepts(
-        cohorts: List[Cohort],
+        cohorts: List[CohortEvaluator],
         top_k: int,
         significance_level: float,
     ) -> Dict[Group, List[TestSummary]]: ...
     @staticmethod
     def calculate_distance_concepts(
-        cohorts: List[Cohort],
+        cohorts: List[CohortEvaluator],
     ) -> Dict[Group, DistanceSummary]: ...
     @staticmethod
     def full_comparison(
-        cohorts: List[Cohort],
+        cohorts: List[CohortEvaluator],
         top_k: int,
         significance_level: float,
     ) -> Tuple[Dict[str, CohortSummary], ComparerSummary]: ...
