+++ medmodels/statistic_evaluations/statistical_analysis/inferential_statistics.pyi
@@ -2,10 +2,13 @@ from typing import List, Tuple
 
 from numpy.typing import ArrayLike
 
-from medmodels.statistic_evaluations.comparer.data_comparer import TestSummary
+from medmodels.medrecord.schema import AttributeType
+from medmodels.statistic_evaluations.evaluate_compare.compare import TestSummary
 
 def normal_distribution_test(sample: ArrayLike) -> bool: ...
-def decide_hypothesis_test(samples: List[ArrayLike], alpha: float) -> TestSummary: ...
+def decide_hypothesis_test(
+    samples: List[ArrayLike], attribute_type: AttributeType, alpha: float
+) -> TestSummary: ...
 def two_tailed_t_test(samples: List[ArrayLike], alpha: float) -> TestSummary: ...
 def mann_whitney_u_test(samples: List[ArrayLike], alpha: float) -> TestSummary: ...
 def analysis_of_variance(samples: List[ArrayLike], alpha: float) -> TestSummary: ...
